Move internal/sliceutil to daemon/internal/sliceutil
