+++ daemon/internal/sliceutil/sliceutil.go
