+++ daemon/libnetwork/internal/resolvconf/resolvconf_test.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/docker/docker/internal/sliceutil"
+	"github.com/docker/docker/daemon/internal/sliceutil"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
