+++ daemon/libnetwork/osl/interface_linux_test.go
@@ -9,8 +9,8 @@ import (
 	"sync"
 	"testing"
 
+	"github.com/docker/docker/daemon/internal/sliceutil"
 	"github.com/docker/docker/daemon/libnetwork/nlwrap"
-	"github.com/docker/docker/internal/sliceutil"
 	"github.com/vishvananda/netlink"
 	"github.com/vishvananda/netns"
 	"gotest.tools/v3/assert"
