+++ daemon/libnetwork/portmappers/routed/mapper_linux.go
@@ -7,9 +7,9 @@ import (
 	"context"
 
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/internal/sliceutil"
 	"github.com/docker/docker/daemon/libnetwork/portmapperapi"
 	"github.com/docker/docker/daemon/libnetwork/types"
-	"github.com/docker/docker/internal/sliceutil"
 )
 
 const driverName = "routed"
