+++ daemon/pkg/plugin/v2/plugin_linux.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"github.com/docker/docker/daemon/internal/rootless/mountopts"
+	"github.com/docker/docker/daemon/internal/sliceutil"
 	"github.com/docker/docker/daemon/pkg/oci"
-	"github.com/docker/docker/internal/sliceutil"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/sys/userns"
 	"github.com/opencontainers/runtime-spec/specs-go"
